pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deskinator"
version = "0.1.0"
description = "Autonomous desk cleaning robot with SLAM"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "EK210 Team"}
]
dependencies = [
    "numpy>=1.26",
    "scipy>=1.13",
    "networkx>=3.2",
    "smbus2>=0.4",
    "matplotlib>=3.7",
]

[project.optional-dependencies]
rpi = [
    "RPi.GPIO; platform_system=='Linux'",
    "adafruit-circuitpython-bno08x>=1.0",
]

[project.scripts]
deskinator = "deskinator.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["deskinator*"]

